requirements.txt
# ArchitectAI — Python dependencies
# Install: pip install -r requirements.txt

# LLM & Agent framework
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0

# IFC generation
ifcopenshell  # install via: pip install ifcopenshell or conda

# DWG parsing
ezdxf>=1.3.0

# PDF parsing
pdfplumber>=0.11.0
pymupdf>=1.24.0

# Vector store & RAG
chromadb>=0.5.0
langchain-chroma>=0.1.0

# Web search (for Compliance Agent self-sourcing)
duckduckgo-search>=6.0.0
requests>=2.32.0
httpx>=0.27.0

# WebSocket server (dashboard live updates)
websockets>=13.0
fastapi>=0.115.0
uvicorn>=0.32.0

# Utilities
pydantic>=2.9.0
python-dotenv>=1.0.0
rich>=13.9.0        # pretty terminal output
loguru>=0.7.0       # structured logging
aiofiles>=24.1.0

# Dev/test
pytest>=8.3.0
pytest-asyncio>=0.24.0
openai>=1.0.0
